requirements.txt
alembic==1.4.2
blinker==1.4
Bootstrap-Flask==1.2.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==7.1.1
coverage==5.0.4
entrypoints==0.3
Faker==4.0.2
flake8==3.7.9
Flask==1.1.2
Flask-CKEditor==0.4.6
Flask-DebugToolbar==0.11.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Moment==0.9.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
idna==3.4
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.2
MarkupSafe==1.1.1
mccabe==0.6.1
pathtools==0.1.2
Pillow==9.4.0
pycodestyle==2.5.0
pyflakes==2.1.1
python-dateutil==2.8.1
python-dotenv==0.12.0
python-editor==1.0.4
requests==2.31.0
six==1.14.0
SQLAlchemy==1.3.15
text-unidecode==1.3
urllib3==2.0.4
watchdog==0.10.2
Werkzeug==1.0.1
WTForms==2.2.1
